US44T89 Move function files into logic folder
