+++ client/src/main/logic/sortingFunctions.ts
